# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MasonPowers Mail API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface RegisterInput {
  /**
   * @minLength 3
   * @maxLength 30
   */
  username: string;
  /** @minLength 6 */
  password: string;
  /** @minLength 1 */
  displayName: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface User {
  id: number;
  username: string;
  email: string;
  displayName: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
}

export interface UsernameAvailability {
  available: boolean;
}

export interface Email {
  id: number;
  fromEmail: string;
  fromName: string;
  toEmail: string;
  toName: string;
  subject: string;
  body: string;
  folder: string;
  isRead: boolean;
  isStarred: boolean;
  isDraft: boolean;
  createdAt: string;
  /** @nullable */
  replyToId?: number | null;
}

export interface EmailListResponse {
  emails: Email[];
  total: number;
  page: number;
  limit: number;
}

export interface EmailInput {
  toEmail: string;
  subject: string;
  body: string;
  isDraft?: boolean;
  /** @nullable */
  replyToId?: number | null;
}

export interface EmailUpdate {
  isRead?: boolean;
  isStarred?: boolean;
  folder?: string;
}

export interface MailboxStats {
  inbox: number;
  sent: number;
  drafts: number;
  trash: number;
  starred: number;
  total: number;
}

export interface Contact {
  email: string;
  displayName: string;
  emailCount: number;
  lastEmailAt: string;
}

export type CheckUsernameParams = {
  username: string;
};

export type ListEmailsParams = {
  folder?: ListEmailsFolder;
  search?: string;
  page?: number;
  limit?: number;
};

export type ListEmailsFolder =
  (typeof ListEmailsFolder)[keyof typeof ListEmailsFolder];

export const ListEmailsFolder = {
  inbox: "inbox",
  sent: "sent",
  drafts: "drafts",
  trash: "trash",
  starred: "starred",
} as const;