# lib/api-zod/src/generated/types/email.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MasonPowers Mail API
 * OpenAPI spec version: 0.1.0
 */

export interface Email {
  id: number;
  fromEmail: string;
  fromName: string;
  toEmail: string;
  toName: string;
  subject: string;
  body: string;
  folder: string;
  isRead: boolean;
  isStarred: boolean;
  isDraft: boolean;
  createdAt: string;
  /** @nullable */
  replyToId?: number | null;
}