# lib/api-zod/src/generated/types/emailInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MasonPowers Mail API
 * OpenAPI spec version: 0.1.0
 */

export interface EmailInput {
  toEmail: string;
  subject: string;
  body: string;
  isDraft?: boolean;
  /** @nullable */
  replyToId?: number | null;
}