# lib/api-zod/src/generated/types/emailListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MasonPowers Mail API
 * OpenAPI spec version: 0.1.0
 */
import type { Email } from "./email";

export interface EmailListResponse {
  emails: Email[];
  total: number;
  page: number;
  limit: number;
}