# lib/api-zod/src/generated/types/emailUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MasonPowers Mail API
 * OpenAPI spec version: 0.1.0
 */

export interface EmailUpdate {
  isRead?: boolean;
  isStarred?: boolean;
  folder?: string;
}