# lib/api-zod/src/generated/types/loginInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MasonPowers Mail API
 * OpenAPI spec version: 0.1.0
 */

export interface LoginInput {
  username: string;
  password: string;
}