# lib/api-zod/src/generated/types/mailboxStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MasonPowers Mail API
 * OpenAPI spec version: 0.1.0
 */

export interface MailboxStats {
  inbox: number;
  sent: number;
  drafts: number;
  trash: number;
  starred: number;
  total: number;
}