# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MasonPowers Mail API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterInput {
  /**
   * @minLength 3
   * @maxLength 30
   */
  username: string;
  /** @minLength 6 */
  password: string;
  /** @minLength 1 */
  displayName: string;
}